Initial perusal vis reducer (#12)
Implement the primary components of Perusal app

Starting to build out the basic architecture for the first version of this app. This should communicate the intended "feel", but it's mainly to establish the major architecture according to my current vision. Subsequent work will be identifying bugs in the current implementation and extending it to cover the full functionality stubbed out now.